Applied suggested items
