+++ cmd/upgrade/status.go
@@ -17,12 +17,7 @@ limitations under the License.
 package upgrade
 
 import (
-	"fmt"
-	"os"
-
-	"github.com/openebs/openebsctl/pkg/upgrade"
 	"github.com/openebs/openebsctl/pkg/upgrade/status"
-	"github.com/openebs/openebsctl/pkg/util"
 	"github.com/spf13/cobra"
 )
 
@@ -32,10 +27,6 @@ func NewCmdUpgradeStatus() *cobra.Command {
 		Aliases: []string{"Status"},
 		Short:   "Display Upgrade-status for a running upgrade-job",
 		Run: func(cmd *cobra.Command, args []string) {
-			if upgrade.CasType != util.JivaCasType {
-				fmt.Println("Only jiva upgrade-status are supported!")
-				os.Exit(1)
-			}
 			status.GetJobStatus()
 		},
 	}
